113-10-22

+++ 113-10-15/Tutorial3_2/Tutorial3_2/Form1.cs
@@ -14,16 +14,16 @@ namespace Tutorial3_2
 
         private void button1_Click(object sender, EventArgs e)
         {
-            double kms;
-            double liters;
-            double average;
+            int kms;
+            int liters;
+            int average;
 
-            kms = double.Parse(txbKm.Text);
-            liters = double.Parse(txbliter.Text);
+            kms = int.Parse(txbKm.Text);
+            liters = int.Parse(txbliter.Text);
             average = kms / liters;
 
 
-            lblshow.Text = (kms/liters).ToString("n3");
+            lblshow.Text = ((double)kms/liters).ToString("n3");
         }
 
         private void btnexit_Click(object sender, EventArgs e)
